feat:add AssignmentMapper
