+++ src/main/java/com/saltpgp/notionproxy/assignment/service/AssignmentService.java
@@ -11,12 +11,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
+import static com.saltpgp.notionproxy.assignment.service.AssignmentMapper.*;
 import static com.saltpgp.notionproxy.service.NotionServiceFilters.filterBuilder;
 
 @Service
 public class AssignmentService {
 
-    public final static String noCommentMessage = "No comment";
     public static final String FILTER = """
             "filter": {
                 "property": "💽 Developer",
@@ -68,31 +68,4 @@ public class AssignmentService {
                 getCategories(properties),
                 getScoreComment(properties));
     }
-
-    private static int getScore(JsonNode properties) {
-        return properties.get("Score").get("number").asInt();
-    }
-
-    private static String getName(JsonNode properties) {
-        return properties.get("Name").get("title").get(0).get("plain_text").asText();
-    }
-
-    private static List<String> getCategories(JsonNode properties) {
-        List<String> categories = new ArrayList<>();
-        if (properties.get("Categories") != null) {
-            properties.get("Categories").get("multi_select").forEach(category ->
-                    categories.add(category.get("name").asText()));
-        }
-        return categories;
-    }
-
-    public static String getScoreComment(JsonNode properties) {
-        try {
-            return properties.get("Comment").get("rich_text")
-                    .get(0).get("plain_text").asText();
-        } catch (Exception e) {
-            return noCommentMessage;
-        }
-    }
-
 }
